[AutoParalllel] Reuse single model for VPP dynamic model (#1240)
